# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateSoftware:
    properties:
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      expiry_date:
        type: string
      install_date:
        type: string
      license_key:
        type: string
      name:
        type: string
      notes:
        type: string
      other_license_info:
        type: string
      server_or_vm:
        type: string
      version:
        type: string
    type: object
...
  models.CreateUser:
    properties:
      contact:
        type: integer
      department:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  models.Software:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      expiry_date:
        type: string
      install_date:
        type: string
      license_key:
        type: string
      name:
        type: string
      notes:
        type: string
      other_license_info:
        type: string
      server_or_vm:
        type: string
      software_id:
        type: integer
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      version:
        type: string
    type: object
...
  models.UpdateSoftware:
    properties:
      customer_location_id:
        type: integer
      expiry_date:
        type: string
      install_date:
        type: string
      license_key:
        type: string
      name:
        type: string
      notes:
        type: string
      other_license_info:
        type: string
      server_or_vm:
        type: string
      updated_by_user_id:
        type: integer
      version:
        type: string
    type: object
...
  models.User:
    properties:
      contact:
        type: integer
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server .
  termsOfService: http://swagger.io/terms/
  title: Network Management Solutions API
  version: "1.0"
paths:
...
  /softwares:
    delete:
      description: Delete Software
      parameters:
      - description: Software ID
        in: query
        name: software_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Software'
      summary: Delete Software
      tags:
      - Software
    get:
      description: Retrieve Software
      parameters:
      - description: Software ID
        in: query
        name: software_id
        type: string
      - description: Select Column
        in: query
        name: select_column
        type: string
      - description: Append Select
        in: query
        name: append_select
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Retrieve  Software
      tags:
      - Software
    patch:
      consumes:
      - application/json
      description: Update Software
      parameters:
      - description: Software
        in: body
        name: Software
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSoftware'
      - description: Software ID
        in: query
        name: software_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Software'
      summary: Update Software
      tags:
      - Software
    post:
      consumes:
      - application/json
      description: Add    Software
      parameters:
      - description: Software Data
        in: body
        name: Software
        required: true
        schema:
          $ref: '#/definitions/models.CreateSoftware'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Software'
      summary: Create Software
      tags:
      - Software
...
  /users/register:
    post:
      consumes:
      - application/json
      description: Add    User
      parameters:
      - description: User Data
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Create User
      tags:
      - User
schemes:
- http
swagger: "2.0"